Toggle ruby postfix conditionals
Using regex search and replace to avoid copying into registers and
possibly overwriting user's data that is already in there. Yay regex!

<leader>I converts three line conditional block to postfix conditional
This places the cursor at the first nonblank character.
User must have the cursor on the line that matches.

<leader>i converts postfix conditional to three line conditional block
This places the cursor on the first line of the nested code.
User should have cursor on the code they want to transform, but side
effect is that this will jump to the previous bit that matches