Remove slow `brew --prefix` calls
While powerlevel10k async management of slow scripts in zsh, it still
causes a brief window in which things are not fully loaded.

This was particulary bad when asdf integration was added to .zprofile
for solargraph/vim/coc.nvim functionality and it did not get the
benefits of the async startup.